requirement.txt
streamlit
langchain
langchain-openai
chromadb
beautifulsoup4
python-dotenv